extract ScalaExpr to separate module
